Changes library to support multiple inputs of different types
This commit changes the interface to the library and its internal
implementation in order to support multiple inputs of different types.
On the interface, the user is now required to provide the number of inputs
and their type when the initialization takes place. Also, when calling
the mp_inc function, the user passes the pointer, the output and then all
the inputs to that call.

On the internal implementation, the library now uses a string key to
represent all inputs. The bit representation is still used for each
input, but it is concatenated into a single string that serves as the key
to that call. When generating the JSON report, this key is printed and can
be easily parsed since all the bit representations are separated by a
'#' character.